refactor: useResetPassword hook
